Add checkered background for opacity.
